modified solr rake task to accept PROFILE parameter
The profile parameter is passed along through getIndex, to specify where
to look for the index generator for a given object.

For AC2, this means that it'll route back to solr/get_index. If it's not
routing to your local dev machine, this means you should be looking
at the dispatch.json config file on lamp.