added facet fields to solr and blacklight config
added the current known facet fields to solr's schema.xml and solrconfig.xml files and to blacklight's blacklight_config.rb. i didn't add 'school' because the details for that field are still being worked out; i did not add ac facet fields to the relevancy weighting fields in solrconfig.xml, and i didn't remove the fields that appear to be scv-specific.

also, a 'title' facet is no longer being considered.