added WIND support, methods to integrate with cul-fedora gem
modified routes in blacklight plugin to make login work (should
investigate in the future what to do this about this in the long
run, examine the default blacklight app)

added application_controller methods to use cul:fedora to get
the fedora/solr servers. (maybe this should be done as
Blacklight.fedora? This seems not right, so for now, just hitches
off application_controller)

added stub for cul-fedora gem